requirements_venv.txt
# CuteCaption Virtual Environment Requirements
# Install with: pip install -r requirements_venv.txt

# Core PyTorch with CUDA support (Windows)
--index-url https://download.pytorch.org/whl/cu118
torch>=2.4.1
torchvision>=0.22.0
torchaudio>=2.7.0

# Core ML/AI libraries
transformers>=4.41.0
accelerate>=0.33.0
bitsandbytes>=0.47.0
safetensors>=0.6.0
faiss-cpu>=1.8.0
git+https://github.com/openai/CLIP.git

# Model-specific dependencies
sentencepiece>=0.2.0
einops>=0.8.0
timm>=1.0.8

# Image processing
Pillow>=11.0.0

# Utilities
tqdm>=4.67.0
pyyaml>=6.0.0
requests>=2.32.0
huggingface-hub>=0.35.0

# System utilities
psutil>=7.1.0

# Note: Mock Triton module is created manually in venv/Lib/site-packages/triton/
# This provides Windows compatibility for CogVLM2 models that require Triton imports
